pyproject.toml
[project]
name = "lczero-training"
version = "0.1.0"
description = "Training scripts and data loading for Leela Chess Zero"
authors = [{ name = "Leela Chess Zero Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pybind11>=2.10.0",
    "numpy>=1.24.0",
    "textual[dev]>=0.47.0",
    "protobuf==6.32.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "anyio>=4.10.0",
    "jax[cuda12]>=0.7.1",
    "flax>=0.11.1",
    "optax>=0.2.5",
    "orbax-checkpoint>=0.11.23",
    "python-dotenv>=1.1.1",
    "requests[socks]>=2.32.5",
    "matplotlib>=3.10.6",
    "tensorboardx>=2.6.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "typing-extensions>=4.0.0",
    "mypy-extensions>=0.4.0",
    "pathspec>=0.11.0",
]

[build-system]
requires = ["setuptools>=64", "pybind11>=2.10.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.so", "*.dll", "*.dylib"]

[dependency-groups]
dev = [
    "mypy-protobuf>=3.6.0",
    "textual-dev>=1.7.0",
    "types-protobuf>=6.30.2.20250809",
    "types-requests>=2.32.4.20250913",
]

[tool.mypy]
mypy_path = "src"

[[tool.mypy.overrides]]
module = "lczero_training.proto.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "orbax.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "optax"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tensorboardX"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["src"]
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["src"]
addopts = "-v"

[tool.ruff]
exclude = ["*_pb2.py*"]
line-length = 80
